Refactored common JS220 & JS110 code to base class.
